# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kayak Trails API",
        "description": "API documentation for kayak trails application. Complete API reference for managing kayak routes, weather data, and geographical information.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api.wartkinurt.pl/api/v1",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/roles": {
            "get": {
                "tags": [
                    "Dashboard - Roles"
                ],
                "summary": "Pobierz listę ról",
                "operationId": "6e047a399347b51b96bedb2d7b798882",
                "parameters": [
                    {
                        "name": "with_permissions",
                        "in": "query",
                        "description": "Dołącz uprawnienia",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "with_users_count",
                        "in": "query",
                        "description": "Dołącz liczbę użytkowników",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista ról pobrana pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dashboard - Roles"
                ],
                "summary": "Utwórz nową rolę",
                "operationId": "0ef9b4ff7da02e946ef6c7bbef03fce5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Manager"
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rola utworzona pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/roles/{role}": {
            "get": {
                "tags": [
                    "Dashboard - Roles"
                ],
                "summary": "Pobierz szczegóły roli",
                "operationId": "8b8675ef53bcac1ce065622f5d95ecb1",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rola pobrana pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dashboard - Roles"
                ],
                "summary": "Zaktualizuj rolę",
                "operationId": "a137c305e455904f3e9130db1ce5c025",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rola zaktualizowana pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dashboard - Roles"
                ],
                "summary": "Usuń rolę",
                "operationId": "7e4225a08dc960e3270734d1baadfd5f",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Rola usunięta pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/roles/{role}/assign-permissions": {
            "post": {
                "tags": [
                    "Dashboard - Roles"
                ],
                "summary": "Przypisz uprawnienia do roli",
                "operationId": "54c30a169841b1d1582626c4bef600ce",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uprawnienia przypisane pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/roles/{role}/revoke-permissions": {
            "delete": {
                "tags": [
                    "Dashboard - Roles"
                ],
                "summary": "Odbierz uprawnienia od roli",
                "operationId": "624ed92b346678b994990bc3794e5ec6",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uprawnienia odebrane pomyślnie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/security/status": {
            "get": {
                "tags": [
                    "Dashboard - System Security"
                ],
                "summary": "Sprawdź status bezpieczeństwa systemu",
                "operationId": "9607d45e2e01dae3a2fdf70b7f2a49f0",
                "responses": {
                    "200": {
                        "description": "Status bezpieczeństwa systemu"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/users/{user}/can-delete": {
            "get": {
                "tags": [
                    "Dashboard - System Security"
                ],
                "summary": "Sprawdź czy użytkownika można bezpiecznie usunąć",
                "operationId": "25f3ff0cfc36896b323d9fafa45e55f0",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informacja o możliwości usunięcia użytkownika"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/security/emergency-info": {
            "get": {
                "tags": [
                    "Dashboard - System Security"
                ],
                "summary": "Pobierz informacje o odzyskiwaniu dostępu w nagłych wypadkach",
                "operationId": "72998970930c7da486b1c3f70fe53b4f",
                "responses": {
                    "200": {
                        "description": "Instrukcje odzyskiwania dostępu"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/security/admin-summary": {
            "get": {
                "tags": [
                    "Dashboard - System Security"
                ],
                "summary": "Pobierz podsumowanie administratorów w systemie",
                "operationId": "e0b3b6bf595759f5a4c6660ac816a93e",
                "responses": {
                    "200": {
                        "description": "Podsumowanie administratorów"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}